api/requirements.txt
celery-pubsub==1.0.0
    celery==5.1.2
        billiard==3.6.4.0
        click==7.1.2
        click-didyoumean==0.0.3
            # click==7.1.2
        click-plugins==1.1.1
            # click==7.1.2
        click-repl==0.2.0
            # click==7.1.2
            prompt-toolkit==3.0.20
                wcwidth==0.2.5
            six==1.16.0
        kombu==5.1.0
            amqp==5.0.6
                vine==5.0.0
            # vine==5.0.0
        pytz==2021.1
        # setuptools
        # vine==5.0.0
django-allow-cidr==0.3.1
    Django==3.2.7
        asgiref==3.4.1
        # pytz==2021.1
        sqlparse==0.4.2
    netaddr==0.8.0
django-celery-beat==2.2.1
    # celery==5.1.2
        # billiard==3.6.4.0
        # click==7.1.2
        # click-didyoumean==0.0.3
            # click==7.1.2
        # click-plugins==1.1.1
            # click==7.1.2
        # click-repl==0.2.0
            # click==7.1.2
            # prompt-toolkit==3.0.20
                # wcwidth==0.2.5
            # six==1.16.0
        # kombu==5.1.0
            # amqp==5.0.6
                # vine==5.0.0
            # vine==5.0.0
        # pytz==2021.1
        # setuptools
        # vine==5.0.0
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
    django-timezone-field==4.2.3
        # Django==3.2.7
            # asgiref==3.4.1
            # pytz==2021.1
            # sqlparse==0.4.2
        # pytz==2021.1
    python-crontab==2.6.0
        python-dateutil==2.8.2
            # six==1.16.0
django-cors-headers==3.8.0
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
django-enumchoicefield==3.0.1
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
django-enumfield==3.1
django-environ==0.7.0
django-extensions==3.1.3
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
django-filter==2.4.0
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
django-mail-factory==0.24
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
    html2text==2020.1.16
django-mailjet==0.3.1
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
    mailjet-rest==1.3.4
        requests==2.26.0
            certifi==2021.5.30
            charset-normalizer==2.0.5
            idna==3.2
            urllib3==1.26.6
django-silk @ git+https://github.com/jazzband/django-silk@7e93d6caecc83f8aad7cdb152b578f27787a59cc
    autopep8==1.5.7
        pycodestyle==2.7.0
        toml==0.10.2
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
    gprof2dot==2021.2.21
    Jinja2==3.0.1
        MarkupSafe==2.0.1
    Pygments==2.10.0
    # python-dateutil==2.8.2
        # six==1.16.0
    # pytz==2021.1
    # requests==2.26.0
        # certifi==2021.5.30
        # charset-normalizer==2.0.5
        # idna==3.2
        # urllib3==1.26.6
    # sqlparse==0.4.2
djangorestframework-api-key==2.2.0
djoser==2.2.2
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
    django-templated-mail==1.1.1
    djangorestframework-simplejwt==5.3.1
        # Django==3.2.7
            # asgiref==3.4.1
            # pytz==2021.1
            # sqlparse==0.4.2
        djangorestframework==3.14.0
            # Django==3.2.7
                # asgiref==3.4.1
                # pytz==2021.1
                # sqlparse==0.4.2
            # pytz==2021.1
        PyJWT==2.8.0
    social-auth-app-django==5.4.0
        # Django==3.2.7
            # asgiref==3.4.1
            # pytz==2021.1
            # sqlparse==0.4.2
        social-auth-core==4.5.1
            cryptography==41.0.7
                cffi==1.16.0
                    pycparser==2.21
            defusedxml==0.8.0rc2
            oauthlib==3.2.2
            # PyJWT==2.8.0
            python3-openid==3.2.0
                # defusedxml==0.8.0rc2
            # requests==2.26.0
                # certifi==2021.5.30
                # charset-normalizer==2.0.5
                # idna==3.2
                # urllib3==1.26.6
            requests-oauthlib==1.3.1
                # oauthlib==3.2.2
                # requests==2.26.0
                    # certifi==2021.5.30
                    # charset-normalizer==2.0.5
                    # idna==3.2
                    # urllib3==1.26.6
drf-extensions==0.7.1
    # djangorestframework==3.14.0
        # Django==3.2.7
            # asgiref==3.4.1
            # pytz==2021.1
            # sqlparse==0.4.2
        # pytz==2021.1
drf-nested-routers==0.93.4
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
    # djangorestframework==3.14.0
        # Django==3.2.7
            # asgiref==3.4.1
            # pytz==2021.1
            # sqlparse==0.4.2
        # pytz==2021.1
drf-spectacular==0.22.1
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
    # djangorestframework==3.14.0
        # Django==3.2.7
            # asgiref==3.4.1
            # pytz==2021.1
            # sqlparse==0.4.2
        # pytz==2021.1
    inflection==0.5.1
    jsonschema==4.6.0
        attrs==21.4.0
        pyrsistent==0.18.1
    PyYAML==6.0
    uritemplate==4.1.1
drf-spectacular-sidecar==2022.6.1
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
drf-yasg==1.20.0
    coreapi==2.3.3
        coreschema==0.0.4
            # Jinja2==3.0.1
                # MarkupSafe==2.0.1
        itypes==1.2.0
        # requests==2.26.0
            # certifi==2021.5.30
            # charset-normalizer==2.0.5
            # idna==3.2
            # urllib3==1.26.6
        # uritemplate==4.1.1
    # coreschema==0.0.4
        # Jinja2==3.0.1
            # MarkupSafe==2.0.1
    # Django==3.2.7
        # asgiref==3.4.1
        # pytz==2021.1
        # sqlparse==0.4.2
    # djangorestframework==3.14.0
        # Django==3.2.7
            # asgiref==3.4.1
            # pytz==2021.1
            # sqlparse==0.4.2
        # pytz==2021.1
    # inflection==0.5.1
    packaging==21.0
        pyparsing==2.4.7
    ruamel.yaml==0.17.16
    # uritemplate==4.1.1
gunicorn==20.1.0
    # setuptools
ordered-set==4.0.2
pipfreeze==2.0.3
psycopg2==2.9.1
python-dotenv==0.19.0
python-slugify==6.1.2
    text-unidecode==1.3
redis==4.1.3
    Deprecated==1.2.13
        wrapt==1.13.3
    # packaging==21.0
        # pyparsing==2.4.7
retry==0.9.2
    decorator==5.1.1
    py==1.11.0
watchdog==2.1.6
whitenoise==5.3.0
